fix(health): address gate review findings
- Fix DiskSpaceCheck status regression: error path could downgrade
  Critical to Warning by adding severity guard
- Fix TODO stubs returning errors that caused permanent StatusCritical
  and StatusUnknown for DetectionRate and DatabasePerformance checks;
  use nil functions with skip guards instead
- Fix log checks returning misleading "healthy" when error buffer has
  no entries; return StatusSkipped when buffer is empty
- Remove no-op MarshalJSON on Result (identity transform adding
  unnecessary allocations)
- Remove dead tmpPath variable in PathAccessCheck
- Fix two consecutive time.Now() calls in ErrorTrendCheck
- Extract shared runChecks helper to deduplicate RunAll/RunCategory
- Move skippedResult helper from audio.go to shared helpers.go
- Fix NotificationProvidersCheck using hardcoded DurationMS: 0
